+++ src/api/controller/torrents.h
@@ -1,11 +1,13 @@
 #ifndef TORREST_TORRENTS_CONTROLLER_H
 #define TORREST_TORRENTS_CONTROLLER_H
 
+#include "boost/filesystem.hpp"
 #include "oatpp/core/macro/codegen.hpp"
 #include "oatpp/core/macro/component.hpp"
 #include "oatpp/web/server/api/ApiController.hpp"
 
 #include "api/dto/file_info_status.h"
+#include "api/dto/items.h"
 #include "api/dto/torrent_info_status.h"
 #include "torrest.h"
 
@@ -30,7 +32,7 @@ public:
 
     ENDPOINT("GET", "/torrents", listTorrents, QUERY(Boolean, status, "status", false)) {
         auto torrents = Torrest::get_instance()->get_service()->get_torrents();
-        auto responseList = oatpp::List<Object<TorrentInfoStatus>>::createShared();
+        auto responseList = List<Object<TorrentInfoStatus>>::createShared();
         for (auto &torrent: torrents) {
             responseList->push_back(status ? TorrentInfoStatus::create(torrent->get_info(), torrent->get_status())
                                            : TorrentInfoStatus::create(torrent->get_info()));
@@ -126,6 +128,97 @@ public:
                                            : FileInfoStatus::create(file->get_info()));
         }
         return createDtoResponse(Status::CODE_200, responseList);
+    }
+
+    ENDPOINT_INFO(torrentItems) {
+        info->summary = "Torrent items";
+        info->description = "Get torrent items";
+        info->pathParams.add<String>("infoHash").description = "Torrent info hash";
+        info->queryParams.add<String>("folder").description = "Items folder";
+        info->queryParams.add<String>("folder").required = false;
+        info->queryParams.add<Boolean>("status").description = "Get files status";
+        info->queryParams.add<Boolean>("status").required = false;
+        info->addResponse<Object<Items>>(Status::CODE_200, "application/json");
+        info->addResponse<Object<ErrorResponse>>(Status::CODE_404, "application/json");
+        info->addResponse<Object<ErrorResponse>>(Status::CODE_500, "application/json");
+    }
+
+    ENDPOINT("GET", "/torrents/{infoHash}/items", torrentItems,
+             PATH(String, infoHash, "infoHash"),
+             QUERY(String, prefix, "folder", ""),
+             QUERY(Boolean, status, "status", false)) {
+        auto fileInfoList = List<Object<FileInfoStatus>>::createShared();
+        auto folderInfoList = List<Object<FolderInfoStatus>>::createShared();
+
+        // Normalize the prefix to remove any trailing slash
+        boost::filesystem::path prefixPath(utils::unescape_string(prefix));
+        if (prefixPath.has_filename() && prefixPath.filename() == ".") {
+            prefixPath.remove_filename();
+        }
+
+        auto files = GET_TORRENT(infoHash)->get_files();
+        for (const auto &file : files) {
+            boost::filesystem::path filePath(file->get_path());
+
+            // Check if filePath starts with prefix
+            auto mismatch = std::mismatch(prefixPath.begin(), prefixPath.end(),
+                                          filePath.begin(), filePath.end());
+
+            // If no mismatch, it means prefix is a leading path in filePath
+            if (mismatch.first == prefixPath.end()) {
+                // Count the number of components in the relative path
+                auto numComponents = std::distance(mismatch.second, filePath.end());
+
+                if (numComponents == 1) {
+                    // If there's only one component, it's a file at the prefix level
+                    fileInfoList->push_back(status ? FileInfoStatus::create(file->get_info(), file->get_status())
+                                                   : FileInfoStatus::create(file->get_info()));
+                } else if (numComponents > 1) {
+                    // If there are more than one component, add the first component as a directory
+                    auto folderName = *mismatch.second;
+                    auto folderPath = prefixPath / folderName;
+
+                    // Build or update our response
+                    auto it = std::find_if(folderInfoList->begin(), folderInfoList->end(),
+                                 [folderPath](const oatpp::data::mapping::type::DTOWrapper<FolderInfoStatus> &f)
+                                           { return f->path == folderPath; });
+
+                    if (it == folderInfoList->end()) {
+                        auto folderInfoStatus = FolderInfoStatus::createShared();
+                        folderInfoStatus->name = folderName.string();
+                        folderInfoStatus->path = folderPath.string();
+                        folderInfoStatus->length = file->get_size();
+                        folderInfoStatus->file_count = 1;
+
+                        if (status) {
+                            auto fileStatus = file->get_status();
+                            folderInfoStatus->status = FolderStatus::createShared();
+                            folderInfoStatus->status->total = fileStatus.total;
+                            folderInfoStatus->status->total_done = fileStatus.total_done;
+                            folderInfoStatus->status->progress = 100.0 * static_cast<double>(folderInfoStatus->status->total_done)
+                                                                 / static_cast<double>(folderInfoStatus->status->total);
+                        }
+
+                        folderInfoList->push_back(folderInfoStatus);
+                    } else {
+                        (*it)->length = (*it)->length + file->get_size();
+                        (*it)->file_count = (*it)->file_count + 1;
+
+                        if (status) {
+                            auto fileStatus = file->get_status();
+                            (*it)->status->total = (*it)->status->total + fileStatus.total;
+                            (*it)->status->total_done = (*it)->status->total_done + fileStatus.total_done;
+                            (*it)->status->progress = 100.0 * static_cast<double>((*it)->status->total_done)
+                                                      / static_cast<double>((*it)->status->total);
+                        }
+                    }
+                }
+            }
+        }
+
+        return (!prefix->empty() && folderInfoList->empty() && fileInfoList->empty())
+                       ? createDtoResponse(Status::CODE_404, ErrorResponse::create("Invalid folder provided"))
+                       : createDtoResponse(Status::CODE_200, Items::create(folderInfoList, fileInfoList));
     }
 
     ENDPOINT_INFO(torrentDownload) {