+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `/torrents/{infoHash}/items` endpoint to allow getting items by folders.
+
 ## [v0.0.6] - 24/09/2024
 
 ### Fixed